moving class LotUpdateDto as an inner class of LotMultiUpdateRequestDto
Issue: IBP-4064
Review: none
