+++ src/main/java/org/generationcp/middleware/domain/inventory/manager/LotMultiUpdateRequestDto.java
@@ -3,15 +3,85 @@ package org.generationcp.middleware.domain.inventory.manager;
 import org.pojomatic.Pojomatic;
 import org.pojomatic.annotations.AutoProperty;
 
+import java.util.LinkedList;
 import java.util.List;
 
 @AutoProperty
 public class LotMultiUpdateRequestDto {
 
-	private List<LotUpdateDto> lotList;
+	@AutoProperty
+	public static class LotUpdateDto {
+
+		private String lotUID;
+
+		private String storageLocationAbbr;
+
+		private String unitName;
+
+		private String notes;
+
+		public LotUpdateDto(){
+
+		}
+
+		public String getLotUID() {
+			return this.lotUID;
+		}
+
+		public void setLotUID(final String lotUID) {
+			this.lotUID = lotUID;
+		}
+
+		public String getStorageLocationAbbr() {
+			return this.storageLocationAbbr;
+		}
+
+		public void setStorageLocationAbbr(final String storageLocationAbbr) {
+			this.storageLocationAbbr = storageLocationAbbr;
+		}
+
+		public String getUnitName() {
+			return this.unitName;
+		}
+
+		public void setUnitName(final String unitName) {
+			this.unitName = unitName;
+		}
+
+		public String getNotes() {
+			return this.notes;
+		}
+
+		public void setNotes(final String notes) {
+			this.notes = notes;
+		}
+
+		@Override
+		public int hashCode() {
+			return Pojomatic.hashCode(this);
+		}
+
+		@Override
+		public String toString() {
+			return Pojomatic.toString(this);
+		}
+
+		@Override
+		public boolean equals(final Object o) {
+			return Pojomatic.equals(this, o);
+		}
+
+	}
+
+
+	private List<LotUpdateDto> lotList = new LinkedList<>();
 
 	public List<LotUpdateDto> getLotList() {
 		return this.lotList;
+	}
+
+	public LotMultiUpdateRequestDto(){
+
 	}
 
 	public void setLotList(final List<LotUpdateDto> lotList) {
@@ -32,4 +102,5 @@ public class LotMultiUpdateRequestDto {
 	public boolean equals(final Object o) {
 		return Pojomatic.equals(this, o);
 	}
+
 }
